refact: unify Eclipse settings across all modules (#974)

+++ org.eclipse.tm4e.ui/src/main/java/org/eclipse/tm4e/ui/internal/utils/MarkerUtils.java
@@ -98,7 +98,7 @@ public final class MarkerUtils {
 				continue;
 			}
 
-			final int lineNumber = lineNumberObj.intValue();
+			final int lineNumber = lineNumberObj;
 			if (lineNumber < startLineNumber) {
 				continue; // this marker needs no update
 			}
